refactor(experimental): fix codecs-data-structures typecheck (#1657)
Currently, running `pnpm test:typecheck` on the `codecs-data-structures` package fails, this PR fixes this by transforming `& object` to `& NonNullable<unknown>` (more famously known as `& {}`) to the `ScalarEnum` type. Umi previously used `& {}` and, when migrating the library to web3.js, I tried to make that type stronger but it seems only `& {}` can make this work.