refactor(experimental): a strategy for expiring a transaction confirmation when a nonce advances (#1700)
# Summary

Transactions with a lifetime tied to a nonce can be thought to expire after that nonce is ‘advanced’. This strategy watches a nonce's account and throws an error if the nonce changes from the expected value (ie. it was advanced in a different transaction).

# Test Plan

```
cd packages/library
pnpm test:unit:browser
pnpm test:unit:node
```