refactor(experimental): a strategy for confirming a transaction when the signature commits (#1701)
# Summary

A transaction can be thought of as ‘confirmed’ when its signature achieves the level of commitment desired. This strategy watches the status of a given signature and returns when the signature's commitment achieves at least the level configured by the confirmer.

# Test Plan

```
cd packages/library
pnpm test:unit:browser
pnpm test:unit:node
```