Create CompilableTransactionMessage (#2457)
This is just a copy of `CompilableTransaction` but in transaction-messages and using the new types

As CompilableTransaction is used in places in transactions that aren't moving to transaction-messages, I've copied instead of moving this

I've also added some type-tests. We had these implicitly in transactions by testing the inputs to partialSign/sign, but these will change to just be `CompilableTransactionMessage`, and there won't be sign functions in transaction-messages. So it makes sense to add some explicit type tests in transaction-messages instead.