Add a decoder + codec for the new transaction type (#2464)
This PR adds a decoder for the new transaction type. It decodes first to an array of signature bytes + an array of message bytes. It then partially decodes the message bytes to identify the signer addresses for the transaction, and uses these to create the signatures map.

It also adds a codec which combines the encoder and decoder.