Refactor compat to convert to new transaction type (#2467)
This PR refactors the compat conversion from old web3js transactions to convert to the new Transaction type.

This is a massive simplification because it no longer needs to decompile the transaction - it just needs to include the serialized message bytes and convert the signatures. We already have helpers to decompile the transaction message if required, so there's no need to duplicate that here.