refactor(experimental): remove program types and resolveTransactionError helper (#2491)
Kinobi no longer makes use of program types such as `Program` and `ProgramWithErrors` as errors are now exported as simple numeric constants which can then be compared with the custom error code provided by the RPC.

As such, the `resolveTransactionError` helper which requires an array of `ProgramWithErrors` instances is also no longer needed. Instead, one can use the `isProgramError` helper and provide the relevant generated constants as illustrated in the previous PR.

This PR removes everything from the `@solana/programs` package aside from the `isProgramError` helper.