Repair the type of the `chain` argument to `useSignTransaction` and `useSignAndSendTransaction` (#2816)
# Summary

When the chains were not known – as in the case of having a plain old `UiWalletAccount` – this invocation would fail to typecheck:

```ts
useSignTransaction(
  account,
  "solana:mainnet" // ERROR
);
```

This was because `OnlySolanaChains<IdentifierArray>` is `never`.

This PR fixes it so that you get both good autocomplete when the chains are known, and you can supply any `solana:*` chain in any case.